# docs/swagger.yaml
basePath: /
definitions:
  types.LoginRequest:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: john_doe
        type: string
    type: object
  types.RegisterRequest:
    properties:
      firstName:
        description: ⬅︎ değişti
        type: string
      lastName:
        description: ⬅︎ değişti
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  types.Restaurant:
    properties:
      description:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
    type: object
  types.User:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password:
        type: string
      role_id:
        type: integer
      role_name:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: RESTful API for restaurant ordering system
  title: SEProject API
  version: "1.0"
paths:
  /admin/restaurant:
    post:
      consumes:
      - application/json
      parameters:
      - description: Yeni restoran
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/types.Restaurant'
      produces:
      - text/plain
      responses:
        "201":
          description: Restoran başarıyla eklendi
          schema:
            type: string
        "400":
          description: Geçersiz veri
          schema:
            type: string
        "500":
          description: Restoran oluşturulamadı
          schema:
            type: string
      summary: Yeni restoran oluştur
      tags:
      - Restaurants
  /admin/restaurant/{id}:
    delete:
      parameters:
      - description: Restoran ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Restoran silindi
          schema:
            type: string
        "400":
          description: Geçersiz ID
          schema:
            type: string
        "500":
          description: Silme başarısız
          schema:
            type: string
      summary: Restoran sil
      tags:
      - Restaurants
    put:
      consumes:
      - application/json
      parameters:
      - description: Restoran ID
        in: path
        name: id
        required: true
        type: integer
      - description: Güncellenmiş restoran
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/types.Restaurant'
      produces:
      - text/plain
      responses:
        "200":
          description: Restoran güncellendi
          schema:
            type: string
        "400":
          description: Geçersiz ID veya veri
          schema:
            type: string
        "500":
          description: Güncelleme başarısız
          schema:
            type: string
      summary: Var olan restoranı güncelle
      tags:
      - Restaurants
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT
      parameters:
      - description: Login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - User
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user with hashed password
      parameters:
      - description: User credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - User
  /restaurants:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Restaurant'
            type: array
        "500":
          description: Restoranlar getirilemedi
          schema:
            type: string
      summary: Tüm restoranları getirir
      tags:
      - Restaurants
  /user:
    get:
      description: Get a user by providing their username
      parameters:
      - description: Username to search
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by username
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Creates a new user in the system
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.User'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "500":
          description: Failed to create user
          schema:
            type: string
      summary: Create a new user
      tags:
      - User
  /user/{id}:
    delete:
      description: Deletes a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
          schema:
            type: string
        "400":
          description: Invalid id
          schema:
            type: string
        "500":
          description: Failed to delete user
          schema:
            type: string
      summary: Delete user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates the user data by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/types.User'
      produces:
      - application/json
      responses:
        "200":
          description: User updated
          schema:
            type: string
        "400":
          description: Invalid id or input
          schema:
            type: string
        "500":
          description: Failed to update user
          schema:
            type: string
      summary: Update user
      tags:
      - User
  /user/all:
    get:
      description: Returns all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users
      tags:
      - User
  /user/userinfo:
    get:
      consumes:
      - application/json
      description: Returns the information of the logged-in user based on their token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get the logged-in user's information
      tags:
      - User
swagger: "2.0"